Fix: rename TdiThreadSafe to TdiThreadStatic to conform filename = Main function convention
