+++ tdishr/TdiDecompileR.c
@@ -35,7 +35,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "tdirefcat.h"
 #include "tdireffunction.h"
 #include "tdirefzone.h"
-#include "tdithreadsafe.h"
+#include "tdithreadstatic.h"
 #include <strroutines.h>
 #include <mds_stdarg.h>
