+++ tdishr/TdiItoX.c
@@ -57,7 +57,7 @@ extern int IsRoprand();
 #include "tdirefcat.h"
 #include "tdirefstandard.h"
 #include "tdinelements.h"
-#include "tdithreadsafe.h"
+#include "tdithreadstatic.h"
 #include <tdishr_messages.h>
 #include <stdlib.h>
 #include <mdsshr.h>
