+++ tdishr/TdiThreadStatic.c
@@ -26,7 +26,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <mdsplus/mdsconfig.h>
 #include <libroutines.h>
 #include <STATICdef.h>
-#include "tdithreadsafe.h"
+#include "tdithreadstatic.h"
 #include <stdlib.h>
 #include <mdsshr.h>
 #include <strroutines.h>
