+++ tdishr/tdithreadstatic.h
